Do code cleanup and abstraction, remove restartMode
Removes restartModeHandlers. This was originally used to restart the
program, but is replaced by a new method to set the state back to ,
found in emitStartOver function. The one concern is that DynamoDB cannot
store empty strings, so if state storage is added this may cause errors.

Abstracts various functions within all handlers. Now the handlers are
used to handle only 'emits' and 'states', while attributes and API
requests are handled by abstracted functions. I may change some of these
functions to listeners (particularly ask and start over) to listeners
themselves.

Adds config.js and resources.js to lambda folder, to separate handler
code from the set up and resources.