Better code cleanup
Moves what were functions into handlers, that are registered along with
the Intent Handlers. This removes the need for Q promises in the
index.js file, and improves the readability a lot too, at least for me.
It's also clearer when things are being emitted as no functions are
being called, only emits.