Remove usage of roomName, update testing and docs
I previously used roomName instead of ownerName, as it made changing
room-names easier superficially. However, it was less obvious
technically, so I entirely replaced roomName with ownerName, so it now
always looks for the actual name of the room calendar, not the name of
the copy created for the Alexa account.

Testing now uses the right duration specified in test-config, and
ownerName throughout.

Docs updated for new changes, and some of the errors corrected. More to
cover still.